move example components to examples folder
