+++ examples/Index.js
@@ -5,22 +5,22 @@ import {
   Example
 } from '@compositor/kit'
 import Page from './Page'
-import Box from './Box'
+import Box from '../src/Box'
 import ExampleBox from './ExampleBox'
-import UtilityBox from './UtilityBox'
-import Flex from './Flex'
-import Heading from './Heading'
-import Link from './Link'
-import Button from './Button'
+import UtilityBox from '../src/UtilityBox'
+import Flex from '../src/Flex'
+import Heading from '../src/Heading'
+import Link from '../src/Link'
+import Button from '../src/Button'
 import theme from './theme'
-import Text from './Text'
-import ButtonPrimary from './ButtonPrimary'
-import ButtonSecondary from './ButtonSecondary'
-import ButtonDanger from './ButtonDanger'
-import Flash from './Flash'
+import Text from '../src/Text'
+import ButtonPrimary from '../src/ButtonPrimary'
+import ButtonSecondary from '../src/ButtonSecondary'
+import ButtonDanger from '../src/ButtonDanger'
+import Flash from '../src/Flash'
 import Octicon from '@github/octicons-react'
-import StateLabel from './StateLabel'
-import CounterLabel from './CounterLabel'
+import StateLabel from '../src/StateLabel'
+import CounterLabel from '../src/CounterLabel'
 
 const Swatch = ({name, index, color, ...rest}) => (
   <div {...rest} key={index}>
