+++ examples/Page.js
@@ -3,7 +3,7 @@ import { ThemeProvider } from 'styled-components'
 import theme from './theme'
 import Meta from './Meta'
 import CSS from './CSS'
-import Box from './Box'
+import Box from '../src/Box'
 
 // Generic page wrapper component
 const Page = props => (
