+++ package.json
@@ -18,8 +18,8 @@
   "license": "MIT",
   "main": "index.js",
   "scripts": {
-    "start": "x0 dev src/Index.js -op 8888",
-    "build": "x0 build src/Index.js --out-dir docs",
+    "start": "x0 dev examples/Index.js -op 8888",
+    "build": "x0 build examples/Index.js --out-dir docs",
     "test": "jest"
   },
   "x0": {
